# spec/openapi.yaml
openapi: 3.1.0
info:
  title: HR Platform API
  version: 1.0.0-alpha.1
  description: |
    Multi-tenant HR platform API.

    **This specification is the contract.** Server interfaces and the Kotlin, Swift and TypeScript
    clients are all generated from it, and a breaking change to a published version fails CI.
    See `docs/03-architecture.md` §9.

    ## Tenancy

    Every request is resolved to exactly one tenant. Precedence, highest first:

    1. The `tenant_id` claim in the bearer token — authoritative once authenticated.
    2. The `X-Tenant-Code` header — for the unauthenticated login and org-resolve endpoints.
    3. The request subdomain.

    Supplying a token and a conflicting `X-Tenant-Code` is rejected with `TENANT_MISMATCH`.

    ## Errors

    Every non-2xx response uses one envelope. Clients localise from the machine-readable `code`,
    never from `message` — the server cannot reliably know the caller's locale, and we ship in
    six languages.

    ## Pagination

    Cursor-based everywhere. Offset pagination is not offered on tenant-scoped collections: it
    re-scans skipped rows and silently drops and duplicates records on concurrently-written
    tables such as attendance and audit.

    ## Idempotency

    Every unsafe method accepts an `Idempotency-Key` header and it is **required** for endpoints
    the mobile outbox can replay. A repeat of a key already processed returns the original
    response rather than applying the operation twice.
...
servers:
...
security:
  - bearerAuth: []
paths:
...
  /v1/directory/search:
    get:
      tags: [Directory]
      summary: Search the employee directory
      description: |
        Available to every authenticated employee — finding a colleague's extension is not a
        privileged operation, and gating it would mean most of the workforce could not use the
        feature they open most often.

        What makes that safe is what the response omits. Salary, bank details, identity documents,
        date of birth and home address are never selected by the query, so there is nothing to
        leak. Anything more sensitive goes through the profile endpoint, which applies
        , a data-scope check and field-level permissions.

        Results are weighted so a name match outranks an employee-code match, which outranks an
        email match. Exited and not-yet-joined employees are excluded.
      operationId: searchDirectory
      parameters:
        - name: query
          in: query
          description: Free text. Tolerates whatever a person types — unbalanced quotes and stray operators do not produce an error.
          schema: { type: string }
        - name: departmentId
          in: query
          schema: { type: string, format: uuid }
        - name: locationId
          in: query
          schema: { type: string, format: uuid }
        - name: cursor
          in: query
          description: Opaque cursor from a previous response.
          schema: { type: string }
        - name: limit
          in: query
          schema: { type: integer, minimum: 1, maximum: 200, default: 50 }
      responses:
        '200':
          description: Matching colleagues
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/DirectoryPage'
  /v1/directory/employees/{id}/reports:
    get:
      tags: [Directory]
      summary: Direct reports of an employee
      description: |
        One level only. The org chart expands a branch at a time rather than fetching an entire
        subtree, which keeps the payload small for a deep organisation and matches how the screen
        is actually used.
      operationId: listDirectReports
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string, format: uuid }
      responses:
        '200':
          description: Direct reports
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/DirectoryEntry'
...
  /v1/employees/me:
    get:
      tags: [Employees]
      summary: The caller's own employee profile
      description: |
        A separate path from `/v1/employees/{id}` rather than making the client substitute its own
        id: the app opens this on launch, before it necessarily knows the employee id.

        Returns `404 NO_EMPLOYEE_RECORD` for a user account that is not linked to an employee —
        a platform operator or an integration credential.
      operationId: getOwnEmployeeProfile
      responses:
        '200':
          description: Profile, filtered to the fields the caller may see
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EmployeeProfile'
        '404':
          $ref: '#/components/responses/NotFound'
  /v1/employees/{id}:
    get:
      tags: [Employees]
      summary: An employee profile
      description: |
        Two independent checks decide what comes back.

        **Whether you see the record at all.** Your own record always; anyone's with
        `employee.view.all` or `employee.manage`; your reporting line — at any depth — with
        `employee.view`. Otherwise `404`, deliberately not `403`: a `403` confirms the record
        exists, which turns this endpoint into an oracle for enumerating employee ids.

        **Which of its fields.** Applied server-side. A field you may not see is **absent from the
        payload**, not null — a null would tell you the field exists and you are not allowed it,
        and would have clients rendering disabled inputs for values they should not know about.
        A masked field arrives redacted, and the true value never leaves the server.

        Sensitive fields — date of birth, home address, personal email, salary grade — are hidden
        by default and require an explicit grant, **including** for callers who hold
        `employee.manage`. Maintaining records and handling identity documents are separable
        duties. Your own sensitive fields are always readable to you.
      operationId: getEmployeeProfile
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string, format: uuid }
      responses:
        '200':
          description: Profile, filtered to the fields the caller may see
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EmployeeProfile'
        '404':
          $ref: '#/components/responses/NotFound'

    patch:
      tags: [Employees]
      summary: Update an employee profile
      description: |
        PATCH rather than PUT because the caller may not be able to read every field. A PUT means
        "here is the whole record", which someone who cannot see the sensitive half is not in a
        position to send — they would have to echo back fields they never received, and the ones
        they omitted would be cleared.

        A field the caller may not write is rejected with `403 FIELD_NOT_WRITABLE`, never silently
        dropped: a save that appears to succeed and quietly discards a change is worse than a
        refusal. Nothing is applied unless everything validates.

        Writing is never the default. Without `employee.manage` you may change only your own
        contact details, preferred name and photo — not your name, date of birth or join date,
        which appear on statutory filings and change by request with evidence.

        Send `If-Match` with the `version` you last read to avoid overwriting a concurrent edit.
      operationId: updateEmployeeProfile
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string, format: uuid }
        - name: If-Match
          in: header
          required: false
          description: |
            The `version` from the profile you loaded, as a weak ETag. Omitted, the update is
            last-write-wins — two HR officers editing the same profile is common enough that
            omitting it loses one of them silently.
          schema: { type: string }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/EmployeeUpdate'
      responses:
        '200':
          description: The updated profile, filtered as on read
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EmployeeProfile'
        '400':
          $ref: '#/components/responses/BadRequest'
        '403':
          $ref: '#/components/responses/Forbidden'
        '404':
          $ref: '#/components/responses/NotFound'
        '409':
          $ref: '#/components/responses/Conflict'
        '422':
          $ref: '#/components/responses/UnprocessableEntity'
  /v1/employees/{id}/form:
    get:
      tags: [Employees]
      summary: The edit form for one employee, with permissions applied
      description: |
        The same schema as `/v1/forms/employee`, narrowed to this caller and this record: fields
        they may not see are absent, and fields they may see but not change arrive with
        `editable: false`.

        Built from the same permission context as the payload and the update path, so what the
        client is offered and what the server will accept cannot drift apart. Deriving the form
        separately is how you end up with an input that saves nothing.
      operationId: getEmployeeEditForm
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string, format: uuid }
        - name: Accept-Language
          in: header
          required: false
          schema: { type: string }
      responses:
        '200':
          description: Form schema for this caller and record
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/FormSchema'
        '404':
          $ref: '#/components/responses/NotFound'
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: Access token from `/v1/auth/token`. Lifetime 15 minutes.
...
  responses:
...
    Forbidden:
      description: Authenticated but not permitted
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ApiErrorResponse'
    NotFound:
      description: Not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ApiErrorResponse'
    BadRequest:
      description: Malformed request, or a field value that could not be interpreted
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ApiErrorResponse'
    Conflict:
      description: |
        The record changed since the caller loaded it. `details` carries `expected` and `actual`
        versions so the client can offer a meaningful choice rather than "please try again".
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ApiErrorResponse'
    UnprocessableEntity:
      description: |
        Syntactically valid but rejected by a domain rule. `details.violations` lists **every**
        problem, not just the first — reporting one at a time turns filling a form into a guessing
        game.
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/ApiErrorResponse'
...
  schemas:
    ApiErrorResponse:
      type: object
      required: [error]
      properties:
        error:
          $ref: '#/components/schemas/ApiError'
    ApiError:
      type: object
      required: [code, message]
      properties:
        code:
          type: string
          description: Stable machine-readable identifier. Clients localise from this.
          examples: [LEAVE_BALANCE_INSUFFICIENT]
        message:
          type: string
          description: Developer-facing English description. Not for display to end users.
        field:
          type: string
          description: Input field the error is attributable to.
        details:
          type: object
          additionalProperties: true
          description: Structured context for building a localised message.
          examples: [{ available: 2.5, requested: 5 }]
        violations:
          type: array
          items:
            $ref: '#/components/schemas/FieldViolation'
        requestId:
          type: string
          description: Correlates with server logs. Quote this to support.
    FieldViolation:
      type: object
      required: [field, code, message]
      properties:
        field: { type: string }
        code: { type: string }
        message: { type: string }
        rejectedValue: {}
...
    FormSchema:
      type: object
      required: [entityType, version, sections]
      properties:
        entityType: { type: string }
        version:
          type: string
          description: |
            A hash of the tenant's field definitions. Cache against it and skip re-rendering when
            unchanged. Deliberately not a timestamp — two servers must agree for the same
            configuration, or a client behind a load balancer sees the schema change on every
            other request.
        sections:
          type: array
          items:
            $ref: '#/components/schemas/FormSection'
    FormSection:
      type: object
      required: [key, label, fields]
      properties:
        key: { type: string }
        label: { type: string }
        fields:
          type: array
          items:
            $ref: '#/components/schemas/FormField'
    FormField:
      type: object
      required: [key, label, type]
      properties:
        key: { type: string }
        label:
          type: string
          description: Already localised. Falls back through region, then English, then the field key — never blank, because a form with an unlabelled field is unusable.
        type:
          type: string
          enum: [TEXT, MULTILINE_TEXT, NUMBER, DATE, DROPDOWN, MULTI_SELECT, RADIO, CHECKBOX, ATTACHMENT, EMPLOYEE, REFERENCE, EMAIL, PHONE]
        required: { type: boolean, default: false }
        editable:
          type: boolean
          default: true
          description: |
            False when the caller may see the value but not change it. Distinct from omitting the
            field — a read-only salary grade still belongs on the profile, it simply is not the
            employee's to edit.
        helpText: { type: string }
        validation:
          $ref: '#/components/schemas/FieldValidation'
        options:
          type: array
          description: For DROPDOWN, MULTI_SELECT and RADIO.
          items:
            $ref: '#/components/schemas/FieldOption'
        referenceTable:
          type: string
          description: |
            For REFERENCE fields: which taxonomy supplies the options, e.g. `nationality`. The
            client already caches these, so the schema names the taxonomy rather than inlining
            several hundred options.
        custom:
          type: boolean
          default: false
          description: True for tenant-defined fields. Clients need not care; support does.
    FieldValidation:
      type: object
      description: |
        Client-side rules, for immediate feedback only. The server re-validates everything —
        anything arriving at the API may have come from a modified app, a replayed request or a
        bulk import.
      properties:
        minLength: { type: integer }
        maxLength: { type: integer }
        min: { type: number }
        max: { type: number }
        pattern:
          type: string
          description: Applied in the client's own regex engine. A malformed pattern is ignored rather than rejecting the user's input — they cannot fix a pattern they cannot see.
        patternMessage: { type: string }
    FieldOption:
      type: object
      required: [value, label]
      properties:
        value: { type: string }
        label: { type: string }
    DirectoryPage:
      type: object
      required: [items, hasMore]
      properties:
        items:
          type: array
          items:
            $ref: '#/components/schemas/DirectoryEntry'
        nextCursor:
          type: string
          description: Pass as `cursor` on the next call. Absent when there are no more results.
        hasMore:
          type: boolean
    DirectoryEntry:
      type: object
      required: [id, employeeCode, displayName]
      description: |
        A colleague, as visible to any authenticated employee.

        Deliberately narrow. Salary, bank details, identity documents, date of birth and home
        address are absent — and absent from the *query*, not filtered afterwards. A serialisation
        filter someone forgets to apply leaks; a column that was never selected cannot.
      properties:
        id: { type: string, format: uuid }
        employeeCode: { type: string }
        displayName:
          type: string
          description: |
            Stored rather than derived from first and last name. Naming order differs by market,
            and several target countries use a mononym or patronymic that does not reconstruct
            from a first/last pair.
        preferredName: { type: string }
        designation: { type: string }
        department: { type: string }
        location: { type: string }
        workEmail: { type: string, format: email }
        mobile: { type: string }
        workPhone: { type: string }
        photoKey:
          type: string
          description: Object storage key. Exchange for a short-lived signed URL to display.
        supervisorId:
          type: string
          format: uuid
          description: The solid reporting line. Approval routing follows this and only this.
    EmployeeProfile:
      type: object
      required: [id, version]
      description: |
        A full employee record, **filtered to what the caller may see**.

        Only `id` and `version` are guaranteed present. Every other property is absent when the
        caller lacks access to it — absent, not null, because a null would confirm the field
        exists and they are not allowed it, and would have clients rendering disabled inputs for
        values they should not know about. Clients must therefore treat every field as optional
        and render from what arrived, never from a fixed list.

        A masked field is present as a redacted string (`••••4567`). The true value never leaves
        the server, so it is not in the HTTP cache, the client database, or any log capturing
        payloads.

        `customFields` holds tenant-defined values, permission-checked individually under their
        own keys and described by `/v1/employees/{id}/form`. The container is omitted entirely
        when nothing in it is visible.
      properties:
        id: { type: string, format: uuid }
        version:
          type: integer
          format: int64
          description: Send back as `If-Match` on update to detect a concurrent edit.
        employeeCode: { type: string }
        status:
          type: string
          enum: [PENDING_JOIN, PROBATION, ACTIVE, SUSPENDED, ON_NOTICE, EXITED]
          description: Read-only here. Changing it is the joiner/leaver process, which has its own approvals and side effects.
        displayName: { type: string }
        firstName: { type: string }
        middleName: { type: string }
        lastName: { type: string }
        preferredName: { type: string }
        photoKey: { type: string }
        titleId: { type: string, format: uuid }
        dateOfBirth: { type: string, format: date }
        genderTypeId: { type: string, format: uuid }
        maritalStatusId: { type: string, format: uuid }
        bloodGroupId: { type: string, format: uuid }
        nationalityId: { type: string, format: uuid }
        religionId: { type: string, format: uuid }
        raceId: { type: string, format: uuid }
        workEmail: { type: string }
        personalEmail: { type: string }
        mobile: { type: string }
        workPhone: { type: string }
        permanentAddress: { $ref: '#/components/schemas/Address' }
        currentAddress: { $ref: '#/components/schemas/Address' }
        companyId: { type: string, format: uuid }
        joinDate: { type: string, format: date }
        confirmationDate: { type: string, format: date }
        probationEndDate: { type: string, format: date }
        resignDate: { type: string, format: date }
        lastWorkingDate: { type: string, format: date }
        employmentTypeId: { type: string, format: uuid }
        employeeCategoryId: { type: string, format: uuid }
        employeeGroupId: { type: string, format: uuid }
        statutoryClassificationId: { type: string, format: uuid }
        departmentId: { type: string, format: uuid }
        designationId: { type: string, format: uuid }
        salaryGradeId: { type: string, format: uuid }
        corporateTitleId: { type: string, format: uuid }
        locationId: { type: string, format: uuid }
        costCentreId: { type: string, format: uuid }
        functionId: { type: string, format: uuid }
        supervisorId:
          type: string
          format: uuid
          description: The solid reporting line. Approval routing follows this and only this.
        dottedLineSupervisorId:
          type: string
          format: uuid
          description: Matrix reporting. Informational — never used for approval routing.
        yearsOfService:
          type: integer
          description: |
            Completed years as at today, stopping at `lastWorkingDate` for a leaver. Drives
            service-based leave slabs and gratuity, both of which are money — hence a server-side
            value rather than something each client computes from `joinDate`.
        customFields:
          type: object
          additionalProperties: true
          description: Tenant-defined values, keyed by field key. Omitted when none are visible.
    EmployeeUpdate:
      type: object
      description: |
        A partial update. Only the properties present are changed; absent ones are untouched.

        To clear a field, name it in `clearFields`. Sending an explicit `null` also works and
        means the same thing, but generated typed clients cannot express it: kotlinx and Swift
        both encode "unset" and "deliberately null" identically, so the encoder must omit nulls
        or every partial update would carry one for each field the caller did not touch.
        `clearFields` is the spelling that survives code generation.

        An empty string also clears a text field — storing `""` would make "not set" and "set to
        nothing" two distinct states that every query downstream then has to handle.

        `status`, `resignDate` and `lastWorkingDate` are absent by design. They are outcomes of
        the joiner/leaver processes, which carry approvals and side effects — payroll cut-off,
        access revocation, final settlement. A profile update that could set `status: EXITED`
        would skip all of it and leave someone paid but locked out, or the reverse.
      properties:
        clearFields:
          type: array
          description: |
            Field keys to set to null. Subject to the same write permissions as setting a value —
            clearing someone's department is a change like any other.

            A key that also appears with a value in the same request is rejected rather than
            resolved by precedence: the two instructions contradict each other, and guessing which
            the caller meant is how a save silently does the opposite of what was intended.
          items: { type: string }
        employeeCode: { type: string, maxLength: 64 }
        firstName: { type: string, maxLength: 128 }
        middleName: { type: [string, "null"], maxLength: 128 }
        lastName: { type: string, maxLength: 128 }
        displayName: { type: string, maxLength: 255 }
        preferredName: { type: [string, "null"], maxLength: 128 }
        photoKey: { type: [string, "null"], maxLength: 512 }
        titleId: { type: [string, "null"], format: uuid }
        dateOfBirth: { type: [string, "null"], format: date }
        genderTypeId: { type: [string, "null"], format: uuid }
        maritalStatusId: { type: [string, "null"], format: uuid }
        bloodGroupId: { type: [string, "null"], format: uuid }
        nationalityId: { type: [string, "null"], format: uuid }
        religionId: { type: [string, "null"], format: uuid }
        raceId: { type: [string, "null"], format: uuid }
        workEmail: { type: [string, "null"], maxLength: 320 }
        personalEmail: { type: [string, "null"], maxLength: 320 }
        mobile: { type: [string, "null"], maxLength: 32 }
        workPhone: { type: [string, "null"], maxLength: 32 }
        permanentAddress: { $ref: '#/components/schemas/Address' }
        currentAddress: { $ref: '#/components/schemas/Address' }
        companyId: { type: string, format: uuid }
        joinDate: { type: string, format: date }
        confirmationDate: { type: [string, "null"], format: date }
        probationEndDate: { type: [string, "null"], format: date }
        employmentTypeId: { type: [string, "null"], format: uuid }
        employeeCategoryId: { type: [string, "null"], format: uuid }
        employeeGroupId: { type: [string, "null"], format: uuid }
        statutoryClassificationId: { type: [string, "null"], format: uuid }
        departmentId: { type: [string, "null"], format: uuid }
        designationId: { type: [string, "null"], format: uuid }
        salaryGradeId: { type: [string, "null"], format: uuid }
        corporateTitleId: { type: [string, "null"], format: uuid }
        locationId: { type: [string, "null"], format: uuid }
        costCentreId: { type: [string, "null"], format: uuid }
        functionId: { type: [string, "null"], format: uuid }
        supervisorId: { type: [string, "null"], format: uuid }
        dottedLineSupervisorId: { type: [string, "null"], format: uuid }
        customFields:
          type: object
          additionalProperties: true
          description: |
            Tenant-defined values. A `null` removes the value. Every key is checked against the
            active definitions — an unknown key is rejected rather than ignored, because silently
            dropping something the user typed is worse than refusing it.
    Address:
      type: object
      description: |
        Free-form because address structure is not universal. A schema with `street`, `city`,
        `state`, `zip` is a US postal address wearing a generic name, and it does not survive
        contact with a Sri Lankan or Japanese address. The keys a tenant uses come from their
        configured address format.
      additionalProperties: true
...